Fixed timing issue
Issues were arriving when communication between serial and the motor
work active at the same time. Found this to be a timing issue and now I
can tell the motor what position to rotate to from the c++ program and
the c++ program will receive data stating what position the motor is
actually achieved.