Overall Improvements
In the arduino code i implemented flush() instead of delays. This made
it to where as soon as the serial write function is done it will move
to the next line of code. This has greatly increased the speed of the
scan. It has also stopped random problems with communication issues.

The midpoint formula based on the averaging of x pixels for each y
pixel has been discarded. The median is now computed which helps to
eliminate the random outlier points.

Using an angle of 65 degrees instead of the measured 33. this was
decided by using libre cad. I used this to measure the angle seen by
the web camera.

All together the scan is much faster and a lot of random outlier points
have been eliminated. Need to figure out the millimeter per pixel at
the distance of the rotating platform to help with the overall geometry
of the shape.