Still fighting the Dynamixel motor
Due to the cheapness of the Dynamixel motor, the motor can not
accurately move to any position unless the next position increment is
>= 25. I am currently implementing an increment of 10. This increment
allows for a change in the motor’s position, even if it is not
precisely what is needed. This will be accounted for in the c++ program.