Done with scanning part of the project
I have finished the scanning side of the code. I can scan objects and
create point clouds based on the object. I am using matlab to test the
accuracy and mesh lab to physically view (it is not as slow as scatter3
function in matlab). The scan seem to be very precise (within a
millimeter precision). Only thing left to do on this side is possibly
adjust the angle measured between the line laser and the camera at the
center of the platform. I will begin to work on the GUI now because it
is not implemented and will come back if time allows.